All Students in a course now display in table on course settings page
Before, the "Students" table would not always display all the users
enrolled in the course. It seemed to happen when there were more than
100 students. Since the table is populated in an unordered fashion,
it would not be clear which students were missing (randomly distributed
through the alphabet). When the List function call was removed for CS60
Spring 2016, more than 100 students were shown. The other List calls for
grutors and instructors were also removed in the case where either of these
were larger than 100 also. These lists were not used elsewhere in this file.